Normalized Q with the current number of policies.
The number of policies made available to the agent in actinfClass
is an arbitrary parameter of the model. It affects the effect of
the prior over precision. However, because in the code the number
of policies changes with time (only those consistent with the
previously-chosen actions are kept, for computer efficiency), the
posteriors over the precision are skewed differently for different
trials, causing an ever-increasing in the sampling value, regardless
of everything else. To fix this, the effect of the number of
policies on the posterior over precision has been eliminated by
normalizing W*Q in the update equations with the number of currently
available policies.