Implemented invp.calculate_Q thingy.
Note that the current implementation assumes that the files
out_123.pi, out_123_0, etc all belong to the same subject. This is
not guaranteed in general. This implementation will be replaced by
one in which invp._save.posteriors() prints the filename that it
actually used to stdout, which will be found in the logs.